requirements.txt
pycryptodome
qrcode
opencv-python
